backend/requirements.txt
Django>=5.0
google-generativeai
requests
gunicorn
psycopg2-binary
django-extensions
twilio
